Serve "static_folders" directory as zip files
To support including download links for my project files on the site and keep them synced with my git repository for each project, the site zips everything in "static_folders" to "static", and copies all "static" files to the build.

This also adds git submodules for XCode projects I want to distribute with the site.